requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.1
cffi==1.9.1
cryptography==1.7.1
Django==1.10.4
django-admin-sortable2==0.6.7
django-celery==3.1.17
django-celery-results==1.0.1
django-jinja==2.2.1
hvac==0.2.17
idna==2.2
Jinja2==2.8
kombu==4.0.2
MarkupSafe==0.23
paramiko==2.1.1
psycopg2==2.6.2
pyasn1==0.1.9
pycparser==2.17
pycrypto==2.6.1
pytz==2016.10
requests==2.12.4
six==1.10.0
slackweb==1.0.5
vine==1.1.3
djangorestframework==3.5.3
django-rest-swagger==2.1.0
django-jet